feat: load listeners and connect to gapi
